fixing evo_info protection and adding tests

+++ src/main/java/pl/psnc/dl/wf4ever/rosrs/Resource.java
@@ -113,10 +113,6 @@ public class Resource {
                 throw new ForbiddenException("Can't update the manifest");
             } else if (researchObject.getFixedEvolutionAnnotationBodyPath().equals(resource)) {
                 throw new ForbiddenException("Can't update the evo info");
-            } else if (ROSRService.SMS.get()
-                    .findAnnotationForBody(researchObject, researchObject.getFixedEvolutionAnnotationBodyPath())
-                    .equals(resource)) {
-                throw new ForbiddenException("Can't delete the evo annotation");
             }
             ResponseBuilder builder = ROSRService.updateInternalResource(researchObject, resource, entity,
                 servletRequest.getContentType());
@@ -316,6 +312,11 @@ public class Resource {
             }
         }
         if (ROSRService.SMS.get().isAnnotation(researchObject, resource)) {
+            if (ROSRService.SMS.get()
+                    .findAnnotationForBody(researchObject, researchObject.getFixedEvolutionAnnotationBodyPath())
+                    .getUri().equals(resource)) {
+                throw new ForbiddenException("Can't delete the evo annotation");
+            }
             return ROSRService.deleteAnnotation(researchObject, resource);
         }
         if (original != null) {
@@ -325,10 +326,6 @@ public class Resource {
             throw new ForbiddenException("Can't delete the manifest");
         } else if (researchObject.getFixedEvolutionAnnotationBodyPath().equals(resource)) {
             throw new ForbiddenException("Can't delete the evo info");
-        } else if (ROSRService.SMS.get()
-                .findAnnotationForBody(researchObject, researchObject.getFixedEvolutionAnnotationBodyPath())
-                .equals(resource)) {
-            throw new ForbiddenException("Can't delete the evo annotation");
         }
 
         return ROSRService.deaggregateInternalResource(researchObject, resource);
